+++ lib/rubocop/ast/traversal.rb
@@ -30,11 +30,11 @@ module RuboCop
                              kwoptarg].freeze
 
       NO_CHILD_NODES.each do |type|
-        module_eval("def on_#{type}(node); end")
+        module_eval("def on_#{type}(node); end", __FILE__, __LINE__)
       end
 
       ONE_CHILD_NODE.each do |type|
-        module_eval(<<-RUBY)
+        module_eval(<<-RUBY, __FILE__, __LINE__ + 1)
           def on_#{type}(node)
             if (child = node.children[0])
               send(:"on_\#{child.type}", child)
@@ -44,7 +44,7 @@ module RuboCop
       end
 
       MANY_CHILD_NODES.each do |type|
-        module_eval(<<-RUBY)
+        module_eval(<<-RUBY, __FILE__, __LINE__ + 1)
           def on_#{type}(node)
             node.children.each { |child| send(:"on_\#{child.type}", child) }
             nil
@@ -54,7 +54,7 @@ module RuboCop
 
       SECOND_CHILD_ONLY.each do |type|
         # Guard clause is for nodes nested within mlhs
-        module_eval(<<-RUBY)
+        module_eval(<<-RUBY, __FILE__, __LINE__ + 1)
           def on_#{type}(node)
             if (child = node.children[1])
               send(:"on_\#{child.type}", child)