+++ lib/rubocop/cop/performance/flat_map.rb
@@ -15,7 +15,7 @@ module RuboCop
       #   [1, 2, 3, 4].map { |e| [e, e] }.flatten
       #   [1, 2, 3, 4].collect { |e| [e, e] }.flatten
       class FlatMap < Cop
-        MSG = 'Use `flat_map` instead of `%s...%s`.'.freeze
+        MSG = 'Use `flat_map` instead of `%<method>s...%<flatten>s`.'.freeze
         FLATTEN_MULTIPLE_LEVELS = ' Beware, `flat_map` only flattens 1 level ' \
                                   'and `flatten` can be used to flatten ' \
                                   'multiple levels.'.freeze
@@ -67,7 +67,8 @@ module RuboCop
 
           add_offense(node,
                       location: range,
-                      message: format(message, first_method, flatten))
+                      message: format(message, method: first_method,
+                                               flatten: flatten))
         end
       end
     end