+++ lib/rubocop/cop/style/empty_else.rb
@@ -6,24 +6,8 @@ module RuboCop
       # Checks for empty else-clauses, possibly including comments and/or an
       # explicit `nil` depending on the EnforcedStyle.
       #
-      # SupportedStyles:
-      #
-      # @example
-      #   # good for all styles
-      #
-      #   if condition
-      #     statement
-      #   else
-      #     statement
-      #   end
-      #
-      #   # good for all styles
-      #   if condition
-      #     statement
-      #   end
-      #
-      # @example
-      #   # empty - warn only on empty else
+      # @example EnforcedStyle: empty
+      #   # warn only on empty else
       #
       #   # bad
       #   if condition
@@ -38,8 +22,20 @@ module RuboCop
       #     nil
       #   end
       #
-      # @example
-      #   # nil - warn on else with nil in it
+      #   # good
+      #   if condition
+      #     statement
+      #   else
+      #     statement
+      #   end
+      #
+      #   # good
+      #   if condition
+      #     statement
+      #   end
+      #
+      # @example EnforcedStyle: nil
+      #   # warn on else with nil in it
       #
       #   # bad
       #   if condition
@@ -54,8 +50,20 @@ module RuboCop
       #   else
       #   end
       #
-      # @example
-      #   # both - warn on empty else and else with nil in it
+      #   # good
+      #   if condition
+      #     statement
+      #   else
+      #     statement
+      #   end
+      #
+      #   # good
+      #   if condition
+      #     statement
+      #   end
+      #
+      # @example EnforcedStyle: both (default)
+      #   # warn on empty else and else with nil in it
       #
       #   # bad
       #   if condition
@@ -68,6 +76,18 @@ module RuboCop
       #   if condition
       #     statement
       #   else
+      #   end
+      #
+      #   # good
+      #   if condition
+      #     statement
+      #   else
+      #     statement
+      #   end
+      #
+      #   # good
+      #   if condition
+      #     statement
       #   end
       class EmptyElse < Cop
         include OnNormalIfUnless